altea: a table's shape follows Signum's — ToStr, row order, column types
Pointed at a Southwind database, the sync asked for ~90 statements that were all
the same kind of thing: altea's declaration of a table disagreed with Signum's
about something small. None of it was a design difference; each was a detail
nobody had had a reason to check.

  • ToStr. A plain `toString()` materialises a `to_str` column; a `@quoted` one is
    expanded inline and materialises nothing. Signum draws the same line with
    [AutoExpressionField] / [ExpressionField], so the two have to agree per type:
    `@quoted` added to ProcessExceptionLine, PackageLine, Case, WorkflowEventTask
    and RestLog, removed from ChatMessage, ResetPasswordRequest, ViewLog and
    TokenEquivalenceGroup.

  • Row order. `@rowOrder` gives a collection table an Order column, which is
    Signum's [PreserveOrder] — and eleven collections carried one where Signum
    does not: a workflow's main-entity strategies are a SET, a lane's actors are
    a SET. Removed where Signum preserves nothing.

  • Column types. `short` and `float` did not exist as value types, so Signum's
    `short` / `float?` columns came out as int / double: added, with `toShort` /
    `toFloat` and the decimal SCALE the four LanguageModel prices need.
    UserQuery's SystemTime window is a date EXPRESSION in Signum (`string?`, not a
    datetime), CachedQuery's two durations are `long`, and four AzureAD / mailing
    ids are `uuid`.

  • Nullability. `@forceNotNullable` is the mirror of `@forceNullable`: the field
    is nullable in the model and the column is not, which is what Signum's
    `string` + no default means for exception.exceptionType and two Help columns.
    Alert.alertDate is the reverse — nullable column, `@notNullValidator`.

  • The columns altea simply lacked: `UserEntity.cultureInfo` (wired into
    EmailLogic's email-owner registration), `UserQuery.createTitle`,
    `TranslatedInstance.rowId` (which also joins its unique index, as Signum's
    does) and `WorkflowScriptEval.customTypes` — carried, not compiled, so a
    Signum application's helper sources are not dropped by a sync.

`typeChanged` compared BOTH dialects' types, so a column matched only when its
SQL Server type matched too — every `long` read as changed on Postgres and the
table was rewritten for nothing. `equalsInDialect` compares the one that is
being generated.

+++ altea-dashboard/data/CachedQuery.ts
@@ -2,7 +2,7 @@ import { reflect, init } from "@altea/altea/data/reflection";
 import { Entity } from "@altea/altea/data/entity";
 import { Lite } from "@altea/altea/data/lite";
 import { entity, backReference, rowOrder, valueField, implementedBy, unit } from "@altea/altea/data/decorators";
-import { Temporal, type int } from "@altea/altea/data/basics";
+import { Temporal, type int, type long } from "@altea/altea/data/basics";
 import { Clock } from "@altea/altea/data/utils/clock";
 import { noRepeatValidator } from "@altea/altea/data/validators";
 import { FilePathEmbedded, FileTypeSymbol } from "@altea/altea-files/data/Files";
@@ -74,10 +74,11 @@ export class CachedQueryEntity extends Entity {
     creationDate: Temporal.PlainDateTime = Clock.now;
 
     @unit("ms")
-    queryDuration: int;
+    // Signum declares both `long` (a millisecond count has no reason to be capped at 24 days).
+    queryDuration: long;
 
     @unit("ms")
-    uploadDuration: int;
+    uploadDuration: long;
 
     // NO toString override, as in Signum: a hand-written one would earn a ToStr column that its table
     // does not have, and the default (the type's nice name plus the id) is what a snapshot row wants
